Create elasticsearch_util and get_elasticsearch_facet()
